Enhance Watchdog Engine and fix World Store loading
- Replaced minimal Watchdog Engine with the full, enhanced version that adheres to AGENTS.md axioms A1-A5.
- Updated playerActions.ts to use the new ValidationResult object from the Watchdog Engine.
- Deleted the redundant watchdog.engine.enhanced.ts file.
- Updated world.store.ts to load metadata from disk synchronously using a dynamic require workaround for ESM environments.
