+++ samples/sample-dev/src/server/hooks/playerActions.ts
@@ -1,6 +1,10 @@
 import { validate } from '../arelogic/watchdog.engine';
 
 export function handleAction(action:any){
- if(!validate(action)) return null;
+ const result = validate(action);
+ if(!result.valid) {
+   if(result.normalized) return result.normalized;
+   return null;
+ }
  return action;
 }
